Separate waveform editor from the additive processor editor.
